+++ hammer-core/src/main/java/cn/featherfly/hammer/tpl/TemplatePreprocessor.java
@@ -1,0 +1,19 @@
+
+package cn.featherfly.hammer.tpl;
+
+/**
+ * PreCompiler.
+ *
+ * @author zhongj
+ */
+@FunctionalInterface
+public interface TemplatePreprocessor {
+
+    /**
+     * Process.
+     *
+     * @param template the template value
+     * @return processed template value
+     */
+    String process(String template);
+}