+++ hammer-core/src/test/java/cn/featherfly/hammer/tpl/mapper/TeMapperImpl.java
@@ -31,7 +31,7 @@ public class TeMapperImpl implements TeMapper {
      */
     @Override
     public User get() {
-        // YUFEI_TODO Auto-generated method stub
+
         return null;
     }
 
@@ -40,7 +40,7 @@ public class TeMapperImpl implements TeMapper {
      */
     @Override
     public User get(String username) {
-        // YUFEI_TODO Auto-generated method stub
+
         return null;
     }
 
@@ -49,7 +49,7 @@ public class TeMapperImpl implements TeMapper {
      */
     @Override
     public User get(String username, String password) {
-        // YUFEI_TODO Auto-generated method stub
+
         return null;
     }
 
@@ -58,7 +58,7 @@ public class TeMapperImpl implements TeMapper {
      */
     @Override
     public User get(String username, String password, Integer age) {
-        // YUFEI_TODO Auto-generated method stub
+
         return null;
     }
 
@@ -67,7 +67,7 @@ public class TeMapperImpl implements TeMapper {
      */
     @Override
     public User get2() {
-        // YUFEI_TODO Auto-generated method stub
+
         return null;
     }
 
@@ -76,7 +76,7 @@ public class TeMapperImpl implements TeMapper {
      */
     @Override
     public User get2(String username, String password, Integer age) {
-        // YUFEI_TODO Auto-generated method stub
+
         return null;
     }
 
@@ -85,7 +85,7 @@ public class TeMapperImpl implements TeMapper {
      */
     @Override
     public Map<String, Object> getMap() {
-        // YUFEI_TODO Auto-generated method stub
+
         return null;
     }
 
@@ -94,7 +94,7 @@ public class TeMapperImpl implements TeMapper {
      */
     @Override
     public Map<String, Object> getMap(String username, String password, Integer age) {
-        // YUFEI_TODO Auto-generated method stub
+
         return null;
     }
 
@@ -103,7 +103,7 @@ public class TeMapperImpl implements TeMapper {
      */
     @Override
     public Map<String, Object> getMap2() {
-        // YUFEI_TODO Auto-generated method stub
+
         return null;
     }
 
@@ -112,7 +112,7 @@ public class TeMapperImpl implements TeMapper {
      */
     @Override
     public Map<String, Object> getMap2(String username, String password, Integer age) {
-        // YUFEI_TODO Auto-generated method stub
+
         return null;
     }
 
@@ -121,7 +121,6 @@ public class TeMapperImpl implements TeMapper {
      */
     @Override
     public void update() {
-        // YUFEI_TODO Auto-generated method stub
 
     }
 
@@ -130,7 +129,6 @@ public class TeMapperImpl implements TeMapper {
      */
     @Override
     public void merge() {
-        // YUFEI_TODO Auto-generated method stub
 
     }
 
@@ -139,7 +137,6 @@ public class TeMapperImpl implements TeMapper {
      */
     @Override
     public void merge(String name, String pass) {
-        // YUFEI_TODO Auto-generated method stub
 
     }
 
@@ -148,7 +145,7 @@ public class TeMapperImpl implements TeMapper {
      */
     @Override
     public int merge2(String name, String pass) {
-        // YUFEI_TODO Auto-generated method stub
+
         return 0;
     }
 
@@ -157,7 +154,7 @@ public class TeMapperImpl implements TeMapper {
      */
     @Override
     public String getName() {
-        // YUFEI_TODO Auto-generated method stub
+
         return null;
     }
 
@@ -166,7 +163,7 @@ public class TeMapperImpl implements TeMapper {
      */
     @Override
     public String getName(Long id) {
-        // YUFEI_TODO Auto-generated method stub
+
         return null;
     }
 
@@ -175,7 +172,7 @@ public class TeMapperImpl implements TeMapper {
      */
     @Override
     public String getName2() {
-        // YUFEI_TODO Auto-generated method stub
+
         return null;
     }
 
@@ -184,7 +181,7 @@ public class TeMapperImpl implements TeMapper {
      */
     @Override
     public String getName2(Long id) {
-        // YUFEI_TODO Auto-generated method stub
+
         return null;
     }
 
@@ -193,7 +190,7 @@ public class TeMapperImpl implements TeMapper {
      */
     @Override
     public int getInt() {
-        // YUFEI_TODO Auto-generated method stub
+
         return 0;
     }
 
@@ -202,7 +199,7 @@ public class TeMapperImpl implements TeMapper {
      */
     @Override
     public int getInt(Long id) {
-        // YUFEI_TODO Auto-generated method stub
+
         return 0;
     }
 
@@ -211,7 +208,7 @@ public class TeMapperImpl implements TeMapper {
      */
     @Override
     public int getInt2() {
-        // YUFEI_TODO Auto-generated method stub
+
         return 0;
     }
 
@@ -220,7 +217,7 @@ public class TeMapperImpl implements TeMapper {
      */
     @Override
     public int getInt2(Long id) {
-        // YUFEI_TODO Auto-generated method stub
+
         return 0;
     }
 
@@ -229,7 +226,7 @@ public class TeMapperImpl implements TeMapper {
      */
     @Override
     public long getLong() {
-        // YUFEI_TODO Auto-generated method stub
+
         return 0;
     }
 
@@ -238,7 +235,7 @@ public class TeMapperImpl implements TeMapper {
      */
     @Override
     public long getLong(Long id) {
-        // YUFEI_TODO Auto-generated method stub
+
         return 0;
     }
 
@@ -247,7 +244,7 @@ public class TeMapperImpl implements TeMapper {
      */
     @Override
     public long getLong2() {
-        // YUFEI_TODO Auto-generated method stub
+
         return 0;
     }
 
@@ -256,7 +253,7 @@ public class TeMapperImpl implements TeMapper {
      */
     @Override
     public long getLong2(Long id) {
-        // YUFEI_TODO Auto-generated method stub
+
         return 0;
     }
 
@@ -265,7 +262,7 @@ public class TeMapperImpl implements TeMapper {
      */
     @Override
     public double getDouble() {
-        // YUFEI_TODO Auto-generated method stub
+
         return 0;
     }
 
@@ -274,7 +271,7 @@ public class TeMapperImpl implements TeMapper {
      */
     @Override
     public double getDouble(Long id) {
-        // YUFEI_TODO Auto-generated method stub
+
         return 0;
     }
 
@@ -283,7 +280,7 @@ public class TeMapperImpl implements TeMapper {
      */
     @Override
     public double getDouble2() {
-        // YUFEI_TODO Auto-generated method stub
+
         return 0;
     }
 
@@ -292,7 +289,7 @@ public class TeMapperImpl implements TeMapper {
      */
     @Override
     public double getDouble2(Long id) {
-        // YUFEI_TODO Auto-generated method stub
+
         return 0;
     }
 
@@ -301,7 +298,7 @@ public class TeMapperImpl implements TeMapper {
      */
     @Override
     public Integer getInteger() {
-        // YUFEI_TODO Auto-generated method stub
+
         return null;
     }
 
@@ -310,7 +307,7 @@ public class TeMapperImpl implements TeMapper {
      */
     @Override
     public Integer getInteger(Long id) {
-        // YUFEI_TODO Auto-generated method stub
+
         return null;
     }
 
@@ -319,7 +316,7 @@ public class TeMapperImpl implements TeMapper {
      */
     @Override
     public Integer getInteger2() {
-        // YUFEI_TODO Auto-generated method stub
+
         return null;
     }
 
@@ -328,7 +325,7 @@ public class TeMapperImpl implements TeMapper {
      */
     @Override
     public Integer getInteger2(Long id) {
-        // YUFEI_TODO Auto-generated method stub
+
         return null;
     }
 
@@ -337,7 +334,7 @@ public class TeMapperImpl implements TeMapper {
      */
     @Override
     public List<User> findUserList() {
-        // YUFEI_TODO Auto-generated method stub
+
         return null;
     }
 
@@ -346,7 +343,7 @@ public class TeMapperImpl implements TeMapper {
      */
     @Override
     public List<User> findUserList(String name) {
-        // YUFEI_TODO Auto-generated method stub
+
         return null;
     }
 
@@ -355,7 +352,7 @@ public class TeMapperImpl implements TeMapper {
      */
     @Override
     public List<User> findUserList2() {
-        // YUFEI_TODO Auto-generated method stub
+
         return null;
     }
 
@@ -364,7 +361,7 @@ public class TeMapperImpl implements TeMapper {
      */
     @Override
     public List<User> findUserList2(String name) {
-        // YUFEI_TODO Auto-generated method stub
+
         return null;
     }
 
@@ -373,7 +370,7 @@ public class TeMapperImpl implements TeMapper {
      */
     @Override
     public List<Map<String, Object>> findMapList() {
-        // YUFEI_TODO Auto-generated method stub
+
         return null;
     }
 
@@ -382,7 +379,7 @@ public class TeMapperImpl implements TeMapper {
      */
     @Override
     public List<Map<String, Object>> findMapList(String name) {
-        // YUFEI_TODO Auto-generated method stub
+
         return null;
     }
 
@@ -391,7 +388,7 @@ public class TeMapperImpl implements TeMapper {
      */
     @Override
     public List<Map<String, Object>> findMapList2() {
-        // YUFEI_TODO Auto-generated method stub
+
         return null;
     }
 
@@ -400,7 +397,7 @@ public class TeMapperImpl implements TeMapper {
      */
     @Override
     public List<Map<String, Object>> findMapList2(String name) {
-        // YUFEI_TODO Auto-generated method stub
+
         return null;
     }
 
@@ -409,7 +406,7 @@ public class TeMapperImpl implements TeMapper {
      */
     @Override
     public PaginationResults<Map<String, Object>> findMapPage(Page page) {
-        // YUFEI_TODO Auto-generated method stub
+
         return null;
     }
 
@@ -418,7 +415,7 @@ public class TeMapperImpl implements TeMapper {
      */
     @Override
     public PaginationResults<Map<String, Object>> findMapPage(String name, Page page) {
-        // YUFEI_TODO Auto-generated method stub
+
         return null;
     }
 
@@ -427,7 +424,7 @@ public class TeMapperImpl implements TeMapper {
      */
     @Override
     public PaginationResults<Map<String, Object>> findMapPage2(Page page) {
-        // YUFEI_TODO Auto-generated method stub
+
         return null;
     }
 
@@ -436,7 +433,7 @@ public class TeMapperImpl implements TeMapper {
      */
     @Override
     public PaginationResults<Map<String, Object>> findMapPage2(String name, Page page) {
-        // YUFEI_TODO Auto-generated method stub
+
         return null;
     }
 
@@ -445,7 +442,7 @@ public class TeMapperImpl implements TeMapper {
      */
     @Override
     public PaginationResults<User> findUserPage(Page page) {
-        // YUFEI_TODO Auto-generated method stub
+
         return null;
     }
 
@@ -454,7 +451,7 @@ public class TeMapperImpl implements TeMapper {
      */
     @Override
     public PaginationResults<User> findUserPage(String name, Page page) {
-        // YUFEI_TODO Auto-generated method stub
+
         return null;
     }
 
@@ -463,7 +460,7 @@ public class TeMapperImpl implements TeMapper {
      */
     @Override
     public PaginationResults<User> findUserPage2(Page page) {
-        // YUFEI_TODO Auto-generated method stub
+
         return null;
     }
 
@@ -472,7 +469,7 @@ public class TeMapperImpl implements TeMapper {
      */
     @Override
     public PaginationResults<User> findUserPage2(String name, Page page) {
-        // YUFEI_TODO Auto-generated method stub
+
         return null;
     }
 
@@ -481,7 +478,7 @@ public class TeMapperImpl implements TeMapper {
      */
     @Override
     public PaginationResults<Map<String, Object>> testPosition(int offset, int limit, String name, String pass) {
-        // YUFEI_TODO Auto-generated method stub
+
         return null;
     }
 
@@ -490,7 +487,7 @@ public class TeMapperImpl implements TeMapper {
      */
     @Override
     public PaginationResults<Map<String, Object>> testPosition(String name, int limit, String pass) {
-        // YUFEI_TODO Auto-generated method stub
+
         return null;
     }
 
@@ -499,7 +496,7 @@ public class TeMapperImpl implements TeMapper {
      */
     @Override
     public PaginationResults<User> testPosition2(int offset, int limit, String name, String pass) {
-        // YUFEI_TODO Auto-generated method stub
+
         return null;
     }
 
@@ -508,7 +505,7 @@ public class TeMapperImpl implements TeMapper {
      */
     @Override
     public PaginationResults<User> testPosition2(String name, int limit, String pass) {
-        // YUFEI_TODO Auto-generated method stub
+
         return null;
     }
 
@@ -517,7 +514,7 @@ public class TeMapperImpl implements TeMapper {
      */
     @Override
     public PaginationResults<User> testPosition2(Page page, String name, String pass) {
-        // YUFEI_TODO Auto-generated method stub
+
         return null;
     }
 
@@ -526,7 +523,7 @@ public class TeMapperImpl implements TeMapper {
      */
     @Override
     public PaginationResults<Map<String, Object>> testPosition(Page page, String name, String pass) {
-        // YUFEI_TODO Auto-generated method stub
+
         return null;
     }
 
@@ -535,7 +532,7 @@ public class TeMapperImpl implements TeMapper {
      */
     @Override
     public List<User> findUserList(Page page) {
-        // YUFEI_TODO Auto-generated method stub
+
         return null;
     }
 
@@ -544,7 +541,7 @@ public class TeMapperImpl implements TeMapper {
      */
     @Override
     public List<User> findUserList2(String name, Page page) {
-        // YUFEI_TODO Auto-generated method stub
+
         return null;
     }
 
@@ -553,7 +550,7 @@ public class TeMapperImpl implements TeMapper {
      */
     @Override
     public List<User> findUserList(int offset, int limit) {
-        // YUFEI_TODO Auto-generated method stub
+
         return null;
     }
 
@@ -562,7 +559,7 @@ public class TeMapperImpl implements TeMapper {
      */
     @Override
     public List<User> findUserList2(String name, int offset, int limit) {
-        // YUFEI_TODO Auto-generated method stub
+
         return null;
     }
 
@@ -571,7 +568,7 @@ public class TeMapperImpl implements TeMapper {
      */
     @Override
     public List<Map<String, Object>> findMapList(Page page) {
-        // YUFEI_TODO Auto-generated method stub
+
         return null;
     }
 
@@ -580,7 +577,7 @@ public class TeMapperImpl implements TeMapper {
      */
     @Override
     public List<Map<String, Object>> findMapList2(String name, Page page) {
-        // YUFEI_TODO Auto-generated method stub
+
         return null;
     }
 
@@ -589,7 +586,7 @@ public class TeMapperImpl implements TeMapper {
      */
     @Override
     public List<Map<String, Object>> findMapList(int offset, int limit) {
-        // YUFEI_TODO Auto-generated method stub
+
         return null;
     }
 
@@ -598,7 +595,7 @@ public class TeMapperImpl implements TeMapper {
      */
     @Override
     public List<Map<String, Object>> findMapList2(String name, int offset, int limit) {
-        // YUFEI_TODO Auto-generated method stub
+
         return null;
     }